moved progress overlay out
